minor improvements

+++ src/main/java/org/codelibs/fess/es/config/exentity/DataConfig.java
@@ -87,7 +87,7 @@ public class DataConfig extends BsDataConfig implements CrawlingConfig {
 
     @Override
     public String getDocumentBoost() {
-        return Float.valueOf(getBoost().floatValue()).toString();
+        return getBoost().toString();
     }
 
     public String getBoostValue() {
